requirements.txt
pytest
requests
pydantic
sqlalchemy
fastapi==0.70.0
psycopg2-binary
uvicorn==0.15.0
websockets
aioredis
alembic
sqlalchemy-utils==0.38.2
trio==0.20.0
asyncpg==0.25.0
pytest-asyncio==0.16.0
httpx==0.20.0
python-jose==3.3.0
anyio==3.6.1
pytest_async==0.1.1
passlib